requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1

# OpenAI + LangChain stack
openai==1.100.2
langchain==0.3.27
langchain-core==0.3.74
langchain-community==0.3.27
langchain-openai==0.3.30
langchain-pinecone==0.2.11
langchain-text-splitters==0.3.9

# Vector DB
pinecone==7.3.0

# PDF ingestion + OCR
PyMuPDF==1.26.3
unstructured[pdf]>=0.14.0
pdfplumber>=0.11.4

# Hybrid search
rank_bm25>=0.2.2
tqdm==4.67.1

# Explicit runtime deps
pydantic==2.11.7
neo4j==5.25.0
psycopg[binary]==3.2.9
psycopg_pool==3.2.4
openpyxl==3.1.5

# S3 + HTTP cliente (opcionales pero recomendados si usarás adjuntos/medios)
boto3>=1.28.0
httpx>=0.27.0
